Issue [#2102655]: Installed an S3-optimized version of image_style_deliver()
Using the core Image module's image_style_deliver() with S3 is slower than
desirable, because it performs an unnecessary extra round trip to S3 due to the
way it serves up the actual image data.

With this new feature, after creating a new image derivative, Drupal issues a 302
Redirect to the file in S3, rather than serving it with file_transfer().