BUGFIX: getimagesize() should no longer fail on certain valid images.
I implemented a way for the mechanism that getimagesize() uses to work for all
images. It was failing on images with large amounts of metadata, due to s3fs's
stream wrapper not fully supporting fseek().

This is still an experimental change, though. I'm not sure that the way I fixed
this is entirely viable in all situations.