Issue #2295157: Restrict collation workaround to MySQL databases.
Due to a bug in Drupal's MySQL driver, hook_schema() implementations cannot
set the collation for a database table directly, which is why s3fs did it
using an ALTER TABLE query. User geodaniel pointed out that this is both
unnecessary and error-causing when using a PostgreSQL database. He provided
a patch that ensures that the ALTER TABLE query is run only on MySQL DBs.