BUGFIX: Calls made to libraries_load() now get around its caching issues.
The awssdk2 library will no longer fail to load if it is actually set up
correctly. If a load failure is detected, the Libraries cache will be cleared
for awssdk2, and another attempt to load it will be made. This gets around
Libraries' insistence that load failures get permacached, which has led to much
frustration for users.