Issue #2361943: Creating image derivatives should no longer sometimes fail.
I never got around to fixing this problem before, because I'd never run into it
myself, and no one had reported it, so it was just theoretical. But it turns out
it IS a good idea to confirm that your derivative actually exists before
redirecting to it. Who'da thunk, right?