Reorganized the configuration page and refactored the code that builds it.
All the settings that aren't absolutely required for s3fs to run are now in the
"Advanced Settings" fieldset, which is collapsed by default. The code is
refactored for readability with that Advanced section in mind.

I also refactored away the S3fsStreamWrapper::cacheControlHeader value. I can't
even remember why I created that in the first place, since
S3fsStreamWrapper::config already exists.