Issue #2574501: Re-engineered the module configuration to better handle overrides in settings.php
Now, when you code a setting into settings.php, that setting will be locked in,
and the form field for it will be disabled. This makes it easier to store your
site configuration in version control, and also makes it clearer from the UI side
when a setting has been hard-coded.